backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic-settings==2.2.1
SQLAlchemy==2.0.30
psycopg[binary]==3.1.18